Make char to be unsigned char by default
OS2EMX_PLAIN_CHAR makes char to be default. As a result, BYTE is typedef
to char. However, BYTE was used as unsigned char. Due to this, bits
operations do not work as expected.

To avoid this, make char to be unsigned char by default. With this,
explicit bits-masking is not necessary.

    modified:   IconMgr.cpp
    modified:   Makefile